dev: Se corrige el menú de Cursos. Se agrega la funcionalidad de modificar un curso; se agrega sobrecarga de operador en clase Curso

+++ src/Utilidades.cpp
@@ -4,7 +4,7 @@ std::string Utilidades::aMinusculas(const std::string& inputUsuario)
 {
 
     std::string resultado = inputUsuario;
-    for(int i = 0; i < resultado.length(); i++)
+    for(int i = 0; i < (int)resultado.length(); i++)
     {
         if(resultado[i] >= 'A' && resultado[i] <= 'Z')
         {
@@ -20,7 +20,7 @@ bool Utilidades::esEnteroValido(const std::string& inputUsuario)
     // funcion para validar que recibamos valores enteros validos
     if(inputUsuario.empty()) return false;
 
-    for (int i = 0; i < inputUsuario.length(); i++)
+    for (int i = 0; i < (int)inputUsuario.length(); i++)
     {
         // se compara con el codigo ASCII
         if (inputUsuario[i] < '0' || inputUsuario[i] > '9') return false;
